+++ server/src/weaverbird/pipeline/steps/sort.py
@@ -3,6 +3,7 @@ from typing import Literal
 from pydantic import BaseModel
 
 from weaverbird.pipeline.steps.utils.base import BaseStep
+from weaverbird.pipeline.steps.utils.render_variables import StepWithVariablesMixin
 from weaverbird.pipeline.types import ColumnName
 
 
@@ -14,3 +15,6 @@ class ColumnSort(BaseModel):
 class SortStep(BaseStep):
     name: Literal["sort"] = "sort"
     columns: list[ColumnSort]
+
+
+class SortStepWithVariables(SortStep, StepWithVariablesMixin): ...