+++ server/tests/steps/test_sort.py
@@ -6,7 +6,8 @@ import pytest
 
 from tests.utils import assert_dataframes_equals
 from weaverbird.backends.pandas_executor.steps.sort import execute_sort
-from weaverbird.pipeline.steps.sort import ColumnSort, SortStep
+from weaverbird.pipeline.steps.sort import ColumnSort, SortStep, SortStepWithVariables
+from weaverbird.utils.toucan_connectors import nosql_apply_parameters_to_query_with_errors
 
 
 @pytest.fixture
@@ -57,3 +58,24 @@ def test_benchmark_sort(benchmark):
         ],
     )
     benchmark(execute_sort, step, df)
+
+
+def test_render_sort_with_variables() -> None:
+    step = SortStepWithVariables(
+        name="sort",
+        columns=[
+            ColumnSort(column="{{ __front_var_0__ }}", order="asc"),
+            ColumnSort(column="{{ __front_var_1__ }}", order="desc"),
+        ],
+    )
+    rendered = step.render(
+        variables={"__front_var_0__": "Company", "__front_var_1__": "Value"},
+        renderer=nosql_apply_parameters_to_query_with_errors,
+    )
+    assert rendered == SortStep(
+        name="sort",
+        columns=[
+            ColumnSort(column="Company", order="asc"),
+            ColumnSort(column="Value", order="desc"),
+        ],
+    )